get rid of useless emits
